final adjustements to Test.Arbitrary
